move command_output to shrs_lang
