+++ shrs_core/src/shell.rs
@@ -86,41 +86,3 @@ pub fn dummy_child() -> anyhow::Result<Child> {
     let cmd = Command::new("true").spawn()?;
     Ok(cmd)
 }
-
-/// Small wrapper that outputs command output if exists
-pub fn command_output(
-    sh: &Shell,
-    ctx: &mut Context,
-    rt: &mut Runtime,
-    cmd_handle: &mut Child,
-) -> anyhow::Result<ExitStatus> {
-    // TODO also handle stderr
-    let output = if let Some(out) = cmd_handle.stdout.take() {
-        let reader = BufReader::new(out);
-        reader
-            .lines()
-            .map(|line| {
-                let line = line.unwrap();
-                println!("{}", line);
-                line
-            })
-            .collect::<Vec<_>>()
-            .join("\n")
-    } else {
-        String::new()
-    };
-
-    // Fetch output status
-    let exit_status = cmd_handle.wait().unwrap().code().unwrap();
-    rt.exit_status = exit_status;
-
-    // Call hook
-    let hook_ctx = AfterCommandCtx {
-        exit_code: exit_status,
-        cmd_time: 0.0,
-        cmd_output: output,
-    };
-    sh.hooks.run::<AfterCommandCtx>(sh, ctx, rt, hook_ctx)?;
-
-    Ok(ExitStatus(exit_status))
-}
